# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "用户中心服务 API 文档",
        "title": "User Hub API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "paths": {
...
        "/api/v1/user-hub/auth/logout": {
            "post": {
                "description": "用户请求吊销其当前的认证令牌（通常是 Refresh Token），使其失效。客户端应在调用此接口后清除本地存储的令牌。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证管理 (Auth Management)"
                ],
                "summary": "退出登录",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"Bearer eyJhbGciOiJI...\"",
                        "description": "Bearer \u003c需要吊销的令牌\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "退出登录成功",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIEmptyResponse"
                        }
                    },
                    "400": {
                        "description": "请求格式错误 (如缺少 Authorization 头或格式非 Bearer)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    },
                    "401": {
                        "description": "认证失败 (通常由 AuthMiddleware 处理，此接口本身逻辑较少触发)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    },
                    "500": {
                        "description": "系统内部错误 (如 Redis 操作失败)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    }
                }
            }
        },
        "/api/v1/user-hub/auth/refresh-token": {
            "post": {
                "description": "使用有效的 Refresh Token 获取一对新的 Access Token 和 Refresh Token。支持从请求体或 Cookie 中获取 Refresh Token。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证管理 (Auth Management)"
                ],
                "summary": "刷新令牌",
                "parameters": [
                    {
                        "description": "请求体 (可选)，包含 refresh_token 字段",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刷新成功，返回新的令牌对",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPITokenPairResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误 (如未提供有效的 Refresh Token)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    },
                    "401": {
                        "description": "认证失败 (Refresh Token 无效、已过期、已被吊销或用户状态异常)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    },
                    "500": {
                        "description": "系统内部错误 (如数据库操作失败、令牌生成失败、Redis 操作失败)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user-hub/users/query": {
            "post": {
                "description": "管理员根据指定的过滤、排序和分页条件，查询用户列表及其关联的 Profile 信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户查询 (User Query)"
                ],
                "summary": "分页查询用户及其资料 (管理员)",
                "parameters": [
                    {
                        "description": "查询条件 (过滤、排序、分页)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserQueryDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功，返回用户列表和总记录数",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIUserListResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数无效 (如JSON格式错误、分页参数超出范围)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    },
                    "403": {
                        "description": "权限不足 (非管理员操作)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    },
                    "500": {
                        "description": "系统内部错误 (如数据库查询失败)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "docs.SwaggerAPIEmptyResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码，0 表示成功，其他值表示错误",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，类型由 T 指定，若无数据则为 nil，且在 JSON 中省略",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.Empty"
                        }
                    ]
                },
                "message": {
                    "description": "可选的响应消息，若为空则不输出",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "docs.SwaggerAPIErrorResponseString": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码，0 表示成功，其他值表示错误",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，类型由 T 指定，若无数据则为 nil，且在 JSON 中省略",
                    "type": "string"
                },
                "message": {
                    "description": "可选的响应消息，若为空则不输出",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "docs.SwaggerAPITokenPairResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码，0 表示成功，其他值表示错误",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，类型由 T 指定，若无数据则为 nil，且在 JSON 中省略",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.TokenPair"
                        }
                    ]
                },
                "message": {
                    "description": "可选的响应消息，若为空则不输出",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "docs.SwaggerAPIUserListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码，0 表示成功，其他值表示错误",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，类型由 T 指定，若无数据则为 nil，且在 JSON 中省略",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.UserListResponse"
                        }
                    ]
                },
                "message": {
                    "description": "可选的响应消息，若为空则不输出",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "dto.RefreshTokenRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "dto.UserQueryDTO": {
            "type": "object",
            "properties": {
                "filters": {
                    "description": "精确匹配条件（如 user_id=\"123\", status=0）",
                    "type": "object",
                    "additionalProperties": true
                },
                "like_filters": {
                    "description": "模糊匹配条件（如 username LIKE \"%test%\"）",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "{\"username\"": " \"test\"}"
                    }
                },
                "order_by": {
                    "description": "排序字段（如 \"created_at DESC\"）",
                    "type": "string",
                    "example": "created_at DESC"
                },
                "page": {
                    "description": "页码，默认 1",
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "page_size": {
                    "description": "每页大小，默认 10",
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1,
                    "example": 10
                },
                "time_range_filters": {
                    "description": "时间范围条件（如 created_at 在某个范围内）",
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "enums.Gender": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-comments": {
                "Female": "女性",
                "Male": "男性",
                "Unknown": "未知"
            },
            "x-enum-varnames": [
                "Unknown",
                "Male",
                "Female"
            ]
        },
...
        "enums.UserRole": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-comments": {
                "RoleAdmin": "0 - 管理员，具有最高权限",
                "RoleGuest": "2 - 访客，限制性访问权限",
                "RoleUser": "1 - 普通用户，标准用户角色"
            },
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleUser",
                "RoleGuest"
            ]
        },
        "enums.UserStatus": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-comments": {
                "StatusActive": "活跃，用户可以正常操作",
                "StatusBlacklisted": "拉黑，用户被禁止访问"
            },
            "x-enum-varnames": [
                "StatusActive",
                "StatusBlacklisted"
            ]
        },
...
        "vo.Empty": {
            "type": "object"
        },
...
        "vo.TokenPair": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "新认证令牌",
                    "type": "string"
                },
                "refresh_token": {
                    "description": "新刷新令牌（可选）",
                    "type": "string"
                }
            }
        },
        "vo.UserListResponse": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.UserWithProfileVO"
                    }
                }
            }
        },
...
        "vo.UserWithProfileVO": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "description": "头像 URL",
                    "type": "string",
                    "example": "https://example.com/avatar.jpg"
                },
                "city": {
                    "description": "城市",
                    "type": "string",
                    "example": "深圳"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "gender": {
                    "description": "性别（0=未知, 1=男, 2=女）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.Gender"
                        }
                    ],
                    "example": 1
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string",
                    "example": "小明"
                },
                "province": {
                    "description": "省份",
                    "type": "string",
                    "example": "广东"
                },
                "role": {
                    "description": "用户角色（0=Admin, 1=User, 2=Guest）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.UserRole"
                        }
                    ],
                    "example": 1
                },
                "status": {
                    "description": "用户状态（0=Active, 1=Blacklisted）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.UserStatus"
                        }
                    ],
                    "example": 0
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "user_id": {
                    "description": "用户 ID",
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
...
    }
}